Consolidate fairness utilities
