+++ docs/Plan.md
@@ -464,7 +464,7 @@ Combine 1-4 and the *effective* context limit becomes hardware bandwidth, not mo
     expands each text file with paraphrases in the translator's languages. The
     helper uses `AutoDatasetFilter` and `LicenseInspector` to keep only clean and
     compliant outputs while logging stats via `DatasetLineageManager`. Measure
-    fairness gains by running `CrossLingualFairnessEvaluator` on the dataset
+    fairness gains by running `fairness.CrossLingualFairnessEvaluator` on the dataset
     before and after augmentation—expect the demographic parity gap to shrink
     by at least 5%.
 40c. **Image and audio fairness metrics**: `FairnessEvaluator.evaluate_multimodal()`
@@ -477,7 +477,7 @@ Combine 1-4 and the *effective* context limit becomes hardware bandwidth, not mo
      ``LLM_PARSER_MODEL`` environment variable to customize loading. Calling
      `download_triples(use_llm_parser=True)` saves triples to
     ``*.triples.json`` files for downstream RAG pipelines ([arxiv.org][15]).
-40e. **Fairness feedback loop**: `fairness_feedback.FairnessFeedback` monitors
+40e. **Fairness feedback loop**: `fairness.FairnessFeedback` monitors
      cross-lingual demographic parity and equal opportunity. When the gap of
      either metric exceeds a configurable threshold it adjusts
      `ActiveDataSelector` or dataset weights through reinforcement learning and
@@ -564,11 +564,11 @@ Combine 1-4 and the *effective* context limit becomes hardware bandwidth, not mo
 60. **Adversarial robustness suite**: Generate gradient-based adversarial prompts and measure model degradation. Acceptable drop is <5% accuracy on the evaluation harness.
 61. **Bias-aware dataset filtering**: Add `DatasetBiasDetector` to compute representation metrics and filter skewed samples. Goal is <5% disparity across demographic slices after filtering.
 61a. **Dataset bias mitigation**: `DataBiasMitigator` reweights or filters entries based on these scores. `download_triples()` now applies the mitigator before storing new files.
-61b. **Fairness gap visualizer**: `fairness_visualizer.FairnessVisualizer` plots demographic parity and opportunity gaps. `dataset_summary.py --fairness-report` saves the charts under `docs/datasets/`; they appear in the lineage and memory dashboards for quick inspection.
+61b. **Fairness gap visualizer**: `fairness.FairnessVisualizer` plots demographic parity and opportunity gaps. `dataset_summary.py --fairness-report` saves the charts under `docs/datasets/`; they appear in the lineage and memory dashboards for quick inspection.
 
 61e. **Fairness adaptation pipeline**: `FairnessAdaptationPipeline` streams bias and cognitive load metrics during ingestion and adjusts `ActiveDataSelector` weights. Bias scores are cached and weight updates are vectorised for faster adaptation. `DatasetLineageManager` logs fairness before and after adaptation, showing improved demographic parity on toy datasets.
 
-61c. **Pre-ingest bias and fairness checks**: `StreamingDatasetWatcher.poll_once()` runs `DatasetBiasDetector` and `CrossLingualFairnessEvaluator` on newly discovered datasets before they are ingested. Reports are saved as `pre_ingest_analysis.json`. Example usage lives in `scripts/pre_ingest_pipeline.py`.
+61c. **Pre-ingest bias and fairness checks**: `StreamingDatasetWatcher.poll_once()` runs `DatasetBiasDetector` and `fairness.CrossLingualFairnessEvaluator` on newly discovered datasets before they are ingested. Reports are saved as `pre_ingest_analysis.json`. Example usage lives in `scripts/pre_ingest_pipeline.py`.
 61d. **Parallelized bias analysis**: `compute_word_freq()` now uses multithreading when available, roughly doubling analysis throughput for large datasets.
 
 62. **Federated world-model training**: Train `world_model_rl` across multiple nodes via gradient averaging. Throughput should scale to four nodes with <1.2× single-node time.
