+++ src/streaming_dataset_watcher.py
@@ -12,7 +12,7 @@ import json
 from .dataset_discovery import DiscoveredDataset, _parse_rss, store_datasets
 from .dataset_summarizer import summarize_dataset
 from .dataset_bias_detector import compute_word_freq, bias_score
-from .cross_lingual_fairness import CrossLingualFairnessEvaluator
+from .fairness import CrossLingualFairnessEvaluator
 try:  # pragma: no cover - optional dependency
     from .data_ingest import CrossLingualTranslator
 except Exception:  # pragma: no cover - missing torch
